Changed field from long to double

+++ src/main/java/br/ufrn/trackingframework/Model/AbstractModel.java
@@ -20,9 +20,9 @@ public class AbstractModel {
     @LastModifiedDate
     private Date lastModifiedDate;
 
-    private Long latitude;
+    private Double latitude;
 
-    private Long longitude;
+    private Double longitude;
 
     private Map<String, Object> data;
 
@@ -63,19 +63,19 @@ public class AbstractModel {
         this.lastModifiedDate = lastModifiedDate;
     }
 
-    public Long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
